requirements.txt
# ══════════════════════════════════════════════════════
# IPL Prediction AI Agent - Python Dependencies
# Install: pip install -r requirements.txt
# ══════════════════════════════════════════════════════

# ── Data Collection & Scraping ──
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
lxml>=4.9.0

# ── Data Processing & Analysis ──
pandas>=2.1.0
numpy>=1.25.0
scipy>=1.11.0

# ── Machine Learning ──
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
shap>=0.43.0

# ── Deep Learning (optional, for advanced models) ──
# torch>=2.1.0
# tensorflow>=2.15.0

# ── Model Management ──
mlflow>=2.8.0
joblib>=1.3.0

# ── Visualization ──
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# ── Web Dashboard ──
streamlit>=1.28.0

# ── Web API ──
fastapi>=0.104.0
uvicorn>=0.24.0

# ── Database ──
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# ── Testing ──
pytest>=7.4.0
pytest-cov>=4.1.0

# ── Utilities ──
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
schedule>=1.2.0
